patch: extract GradientSettings to its own module
